requirements.txt
dash
openai
requests
python-dotenv
pillow
tiktoken
